Address feedback from review
