+++ src/Tools/Shared/SecretsHelpers/assets/SecretManager.targets
